dj_backend/requirements.txt
djangorestframework-jwt==1.11.0
djangorestframework==3.11.1
django-cors-middleware==1.5.0
